% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_bsafe.R
\name{bsafe}
\alias{bsafe}
\alias{bsafe_UI}
\alias{bsafe_server}
\title{bsafe's ui function}
\usage{
bsafe_UI(id, header = NULL)

bsafe_server(id, dataset)
}
\arguments{
\item{id}{the id}

\item{header}{elements to be included in the header of the tabset of the module}

\item{dataset}{the dataset to be included in the application}
}
\value{
the UI

the UI
}
\description{
bsafe's ui function
}
\section{Functions}{
\itemize{
\item \code{bsafe_UI()}: bsafe UI function UI for the bsafe shiny app

\item \code{bsafe_server()}: bsafe server function for the bsafe shiny app

}}
