% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bsafe_data}
\alias{bsafe_data}
\title{bsafe dummy data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 30 rows and 12 columns.
}
\usage{
bsafe_data
}
\description{
bsafe dummy data
}
\keyword{datasets}
