% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_bsafe.R
\name{bsafe_UI}
\alias{bsafe_UI}
\title{quic.bsafe's ui function}
\usage{
bsafe_UI(id, header = NULL)
}
\arguments{
\item{id}{the id}

\item{header}{elements to be included in the header of the tabset of the module}
}
\value{
the UI
}
\description{
implements the UI for the quic.bsafe shiny app
}
