% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_data_preparation.R
\name{create_arm}
\alias{create_arm}
\title{New arm creation}
\usage{
create_arm(data, new_arm, col_vals)
}
\arguments{
\item{data}{original data}

\item{new_arm}{new arm name}

\item{col_vals}{filtering criteria to create the new arm}
}
\description{
New arm creation
}
