% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_data}
\alias{test_data}
\title{bsafe test dataset}
\format{
An object of class \code{data.frame} with 23 rows and 8 columns.
}
\usage{
test_data
}
\description{
bsafe test dataset
}
\keyword{datasets}
