% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_dm_in}
\alias{test_dm_in}
\title{test data for decision making}
\format{
An object of class \code{list} of length 10.
}
\usage{
test_dm_in
}
\description{
test data for decision making
}
\keyword{datasets}
