% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_na_in}
\alias{test_na_in}
\title{test data for new analysis trial}
\format{
An object of class \code{list} of length 7.
}
\usage{
test_na_in
}
\description{
test data for new analysis trial
}
\keyword{internal}
