% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_rm_in}
\alias{test_rm_in}
\title{test data for robust map}
\format{
An object of class \code{list} of length 8.
}
\usage{
test_rm_in
}
\description{
test data for robust map
}
\keyword{datasets}
