% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_bsafe.R
\name{tm_bsafe}
\alias{tm_bsafe}
\title{tm_bsafe}
\usage{
tm_bsafe(label = "BSAFE", dataset_name)
}
\arguments{
\item{label}{a label to be used for the application}

\item{dataset_name}{the names of the dataset name}
}
\description{
tm_bsafe
}
